+++ src/algorithm/NoHitException.java
@@ -1,4 +1,4 @@
-package utility;
+package algorithm;
 
 public class NoHitException extends Exception {
     public NoHitException(String message) {
