+++ src/application/Main.java
@@ -1,4 +1,4 @@
-package ui;
+package application;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
@@ -10,7 +10,7 @@ public class Main extends Application {
 
     @Override
     public void start(Stage primaryStage) throws Exception{
-        Parent root = FXMLLoader.load(getClass().getResource("viewer.fxml"));
+        Parent root = FXMLLoader.load(getClass().getResource("view/viewer.fxml"));
         primaryStage.setTitle("IntelliMow Viewer");
         primaryStage.setScene(new Scene(root));
         primaryStage.show();
