+++ src/application/img/icon.ico
