+++ src/application/img/logo-wide.png
