+++ src/application/view/EditorController.java
@@ -1,4 +1,4 @@
-package ui;
+package application.view;
 
 import algorithm.Boundary;
 import algorithm.UIObject;
@@ -9,7 +9,6 @@ import javafx.scene.canvas.Canvas;
 import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.control.*;
 import javafx.scene.input.MouseEvent;
-import javafx.scene.paint.Color;
 import javafx.scene.shape.Circle;
 
 import java.io.File;
